requirements.txt
numpy==1.24.3
tensorflow==2.13.1
tqdm==4.64.1
torch==2.1.1+cu118 
scikit-learn==1.0.2 #1.3.0
opencv_python==4.5.4.60
pillow==10.0.1
pandas==2.1.4
matplotlib==3.8.0
#keras=2.13.1
#Cython==3.0.8